requirements.txt
selenium==3.14.0
chromedriver-binary==86.0.4240.22.0
pygsheets
html-table-parser-python3